China Global Investment Tracker

The China Global Investment Tracker (CGIT) is the only comprehensive public data set covering China’s global investment and construction, which are documented both separately and together. Inaugurated in 2005, the CGIT now includes over 3,300 large transactions across energy, transportation, real estate, and other industries, as well as 300 troubled transactions. The full list, with the amount, Chinese parent company, host country, and sector, is available for public use with the proper citation. The tracker is published by the American Enterprise Institute.

cgit_investments_df

cgit_construction_contracts_df

cgit_troubled_transactions_df

cgit_combined_df

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1635 rows and 12 columns.

Source

https://www.aei.org/china-global-investment-tracker/

References

The American Enterprise Institute and The Heritage Foundation

Examples

cgit_investments_df

#> # A tibble: 1,635 x 12
#>     Year Month Investor `Quantity in Mi… `Share Size` `Transaction Pa… Sector
#>    <dbl> <chr> <chr>               <dbl> <chr>        <chr>            <chr> 
#>  1  2005 Janu… Minmeta…              500 NA           Cubapetroleo     Metals
#>  2  2005 Janu… China A…             1740 NA           IBM              Techn…
#>  3  2005 Febr… Minmeta…              550 0.5          Codelco          Metals
#>  4  2005 March MCC                   670 0.85         Highlands Pacif… Metals
#>  5  2005 April CNOOC                 130 0.17         MEG Energy       Energy
#>  6  2005 June  Sinopec               120 0.4          Synenco          Energy
#>  7  2005 July  Jiangsu…              100 1            MG               Trans…
#>  8  2005 Augu… CNPC                 4200 0.67         PetroKazakhstan  Energy
#>  9  2005 Sept… CNPC, S…             1420 NA           EnCana           Energy
#> 10  2005 Sept… State-O…              200 NA           NA               Energy
#> # … with 1,625 more rows, and 5 more variables: Subsector <chr>, Country <chr>,
#> #   Region <chr>, BRI <dbl>, Greenfield <chr>

cgit_construction_contracts_df

#> # A tibble: 1,700 x 11
#>     Year Month Contractor `Quantity in Mi… `Share Size` `Transaction Pa… Sector
#>    <dbl> <chr> <chr>                 <dbl>        <dbl> <chr>            <chr> 
#>  1  2005 Janu… Sinohydro               100           NA NA               Trans…
#>  2  2005 Febr… Sinohydro               150           NA Oman Wastewater… Utili…
#>  3  2005 Febr… Sinohydro               370           NA NA               Trans…
#>  4  2005 March Sinoma                  170           NA Riyadh Cement    Real …
#>  5  2005 April Sinomach                100           NA NA               Utili…
#>  6  2005 April Minmetals…              240           NA NA               Metals
#>  7  2005 May   CNPC                    390           NA SONATRACH        Energy
#>  8  2005 June  Sinomach,…              860           NA NA               Energy
#>  9  2005 June  MCC                     170           NA NA               Trans…
#> 10  2005 June  China Rai…             1270           NA Cengiz and IC I… Trans…
#> # … with 1,690 more rows, and 4 more variables: Subsector <chr>, Country <chr>,
#> #   Region <chr>, BRI <dbl>

cgit_troubled_transactions_df

#> # A tibble: 290 x 11
#>     Year Month `Chinese Entity` `Quantity in Mi… `Transaction Pa… Sector
#>    <dbl> <chr> <chr>                       <dbl> <chr>            <chr> 
#>  1  2005 July  CNOOC                       18000 Unocal           Energy
#>  2  2006 July  China Mobile                 5300 Millicom         Techn…
#>  3  2006 July  CNPC                         1390 KazMunaiGaz      Energy
#>  4  2006 July  CNPC                         2500 Rosneft          Energy
#>  5  2006 July  China Ocean Shi…              140 Burg             Trans…
#>  6  2006 Sept… CNOOC                       16000 NA               Energy
#>  7  2006 Octo… Southern Power                190 NA               Energy
#>  8  2006 Nove… Chinalco                      790 Vietnam Nationa… Metals
#>  9  2006 Nove… China Railway C…             7450 NA               Trans…
#> 10  2006 Dece… Jinchuan                     1000 Philnico         Metals
#> # … with 280 more rows, and 5 more variables: Subsector <chr>, Country <chr>,
#> #   Region <chr>, Greenfield <chr>, BRI <dbl>

cgit_combined_df

#> # A tibble: 3,335 x 12
#>     Year Month `Chinese Entity` `Quantity in Mi… `Share size` `Transaction Pa…
#>    <dbl> <chr> <chr>                       <dbl> <chr>        <chr>           
#>  1  2005 Janu… Minmetals                     500 NA           Cubapetroleo    
#>  2  2005 Janu… China Academy o…             1740 NA           IBM             
#>  3  2005 Janu… Sinohydro                     100 NA           NA              
#>  4  2005 Febr… Minmetals                     550 0.5          Codelco         
#>  5  2005 Febr… Sinohydro                     150 NA           Oman Wastewater…
#>  6  2005 Febr… Sinohydro                     370 NA           NA              
#>  7  2005 March MCC                           670 0.85         Highlands Pacif…
#>  8  2005 March Sinoma                        170 NA           Riyadh Cement   
#>  9  2005 April CNOOC                         130 0.17         MEG Energy      
#> 10  2005 April Sinomach                      100 NA           NA              
#> # … with 3,325 more rows, and 6 more variables: Sector <chr>, Subsector <chr>,
#> #   Country <chr>, Region <chr>, BRI <lgl>, Greenfield <chr>