World Bank's World Development Indicators

The World Development Indicators is a compilation of relevant, high-quality, and internationally comparable statistics about global development and the fight against poverty. The database contains 1,600 time series indicators for 217 economies and more than 40 country groups, with data for many indicators going back more than 50 years.

world_bank_df

Format

An object of class tbl_df (inherits from tbl, data.frame) with 7274562 rows and 6 columns.

Source

World Bank

Examples

world_bank_df

#> # A tibble: 7,274,562 x 6
#>    country_name country_code indicator_name         indicator_code   year  value
#>    <fct>        <fct>        <fct>                  <fct>           <int>  <dbl>
#>  1 Arab World   ARB          Adolescent fertility … SP.ADO.TFRT      1960 1.35e2
#>  2 Arab World   ARB          Age dependency ratio … SP.POP.DPND      1960 8.81e1
#>  3 Arab World   ARB          Age dependency ratio,… SP.POP.DPND.OL   1960 6.59e0
#>  4 Arab World   ARB          Age dependency ratio,… SP.POP.DPND.YG   1960 8.13e1
#>  5 Arab World   ARB          Aquaculture productio… ER.FSH.AQUA.MT   1960 4.60e3
#>  6 Arab World   ARB          Arms exports (SIPRI t… MS.MIL.XPRT.KD   1960 3.00e6
#>  7 Arab World   ARB          Arms imports (SIPRI t… MS.MIL.MPRT.KD   1960 5.39e8
#>  8 Arab World   ARB          Birth rate, crude (pe… SP.DYN.CBRT.IN   1960 4.78e1
#>  9 Arab World   ARB          Broad money (% of GDP) FM.LBL.BMNY.GD…  1960 2.21e1
#> 10 Arab World   ARB          Capture fisheries pro… ER.FSH.CAPT.MT   1960 4.75e5
#> # … with 7,274,552 more rows